session/requirements.txt
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.119
botocore==1.34.119
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dj-rest-auth==6.0.0
Django==5.0.3
django-allauth==0.63.2
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.13.0
elasticsearch==8.13.0
filelock==3.13.4
fsspec==2024.3.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.22.2
idna==3.6
inflection==0.5.1
jmespath==1.0.1
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
s3transfer==0.10.1
safetensors==0.4.3
selenium==4.20.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tokenizers==0.19.1
tqdm==4.66.2
transformers==4.40.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
